warning più intelligenti

+++ src/it/enne/curie/common/LogWriter.java
@@ -11,12 +11,12 @@ public class LogWriter {
     private int rowCont;
 
     public LogWriter(String logPath) {
+        boolean isCreated;
         log = new File(logPath);
 
         try {
             // cartella controllo
             File cartella = new File(getFolderName());
-            boolean isCreated;
             if (!cartella.exists() || !cartella.isFile()) {
                 isCreated = cartella.mkdirs();
                 if (isCreated) {
@@ -49,40 +49,42 @@ public class LogWriter {
     }
 
     public void write(String line) {
-        boolean isWritable = log.setWritable(true, true);
-        if (!isWritable) {
-            System.err.println("errore scrittura log");
-        }
+        boolean isWritable;
 
-        if (rowCont>=20) { // eliminazione prima riga file
-            File temp = new File(log+"tmp");
-            try (
-                    PrintWriter writer = new PrintWriter(new BufferedWriter(new OutputStreamWriter(new FileOutputStream(temp))));
-                    BufferedReader br = new BufferedReader(new FileReader(log))
-            ) {
-                String let;
-                br.readLine();
-                while ((let = br.readLine()) != null) {
-                    writer.write(let+"\n");
+        isWritable = log.setWritable(true,true);
+        if (isWritable) {
+            if (rowCont >= 20) { // eliminazione prima riga file
+                File temp = new File(log + "tmp");
+                try (
+                        PrintWriter writer = new PrintWriter(new BufferedWriter(new OutputStreamWriter(new FileOutputStream(temp))));
+                        BufferedReader br = new BufferedReader(new FileReader(log))
+                ) {
+                    String let;
+                    br.readLine();
+                    while ((let = br.readLine()) != null) {
+                        writer.write(let + "\n");
+                    }
+                    rowCont--;
+                } catch (Exception e) {
+                    System.err.println("Errore eliminazione prima riga file log");
                 }
-                rowCont--;
+                if (!log.delete() || !temp.renameTo(log)) {
+                    System.err.println("errore switch  file temporaneo file log");
+                }
+            }
+            // aggiunta riga in fondo al file di log
+            try (PrintWriter writer = new PrintWriter(new BufferedWriter(new OutputStreamWriter(new FileOutputStream(log, true))))) {
+                writer.append(line).append("\n");
+                rowCont++;
             } catch (Exception e) {
-                System.err.println("Errore eliminazione prima riga file log");
+                System.err.println("Errore scrittura nel log: " + line);
             }
-            if (!log.delete() || !temp.renameTo(log)) {
-                System.err.println("errore switch  file temporaneo file log");
+            isWritable = log.setWritable(false, false);
+            if (!isWritable) {
+                System.err.println("log modificabile");
             }
-        }
-        // aggiunta riga in fondo al file di log
-        try (PrintWriter writer = new PrintWriter(new BufferedWriter(new OutputStreamWriter(new FileOutputStream(log, true))))) {
-            writer.append(line).append("\n");
-            rowCont++;
-        } catch (Exception e) {
-            System.err.println("Errore scrittura nel log: " + line);
-        }
-        isWritable = log.setWritable(false,false);
-        if (!isWritable) {
-            System.err.println("log modificabile");
+        } else {
+            System.err.println("errore scrittura log");
         }
     }
